drone/pyproject.toml
[project]
name = "drone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0",
    "paho-mqtt>=2.1.0",
    "stmpy>=0.7.0",
]
